add test for other matcher
