+++ packages/expect-puppeteer/src/matchers/notToMatch.test.js
@@ -1,9 +1,15 @@
+import { setupPage } from './setupPage'
+
 describe('not.toMatch', () => {
   beforeEach(async () => {
     await page.goto(`http://localhost:${process.env.TEST_SERVER_PORT}`)
   })
 
-  describe('Page', () => {
+  describe.each(['Page', 'Frame'])('%s', pageType => {
+    let page
+    setupPage(pageType, ({ currentPage }) => {
+      page = currentPage
+    })
     it('should be ok if text is not in the page', async () => {
       await expect(page).not.toMatch('Nop!')
     })
