+++ packages/expect-puppeteer/src/matchers/notToMatchElement.test.js
@@ -1,9 +1,15 @@
+import { setupPage } from './setupPage'
+
 describe('not.toMatchElement', () => {
   beforeEach(async () => {
     await page.goto(`http://localhost:${process.env.TEST_SERVER_PORT}`)
   })
 
-  describe('Page', () => {
+  describe.each(['Page', 'Frame'])('%s', pageType => {
+    let page
+    setupPage(pageType, ({ currentPage }) => {
+      page = currentPage
+    })
     it('should not match using selector', async () => {
       await expect(page).not.toMatchElement('wtf')
     })
