+++ packages/expect-puppeteer/src/matchers/toClick.test.js
@@ -1,9 +1,15 @@
+import { setupPage } from './setupPage'
+
 describe('toClick', () => {
   beforeEach(async () => {
     await page.goto(`http://localhost:${process.env.TEST_SERVER_PORT}`)
   })
 
-  describe('Page', () => {
+  describe.each(['Page', 'Frame'])('%s', pageType => {
+    let page
+    setupPage(pageType, ({ currentPage }) => {
+      page = currentPage
+    })
     it('should click using selector', async () => {
       await expect(page).toClick('a[href="/page2.html"]')
       await page.waitForSelector('html')
