+++ packages/expect-puppeteer/src/matchers/toSelect.test.js
@@ -1,9 +1,15 @@
+import { setupPage } from './setupPage'
+
 describe('toSelect', () => {
   beforeEach(async () => {
     await page.goto(`http://localhost:${process.env.TEST_SERVER_PORT}`)
   })
 
-  describe('Page', () => {
+  describe.each(['Page', 'Frame'])('%s', pageType => {
+    let page
+    setupPage(pageType, ({ currentPage }) => {
+      page = currentPage
+    })
     it('should select an option using value', async () => {
       await expect(page).toSelect('select[name="my-select"]', 'opt1')
       const currentValue = await page.evaluate(
