+++ packages/expect-puppeteer/src/matchers/toUploadFile.test.js
@@ -1,11 +1,16 @@
 import path from 'path'
+import { setupPage } from './setupPage'
 
 describe('toUploadFile', () => {
   beforeEach(async () => {
     await page.goto(`http://localhost:${process.env.TEST_SERVER_PORT}`)
   })
 
-  describe('Page', () => {
+  describe.each(['Page', 'Frame'])('%s', pageType => {
+    let page
+    setupPage(pageType, ({ currentPage }) => {
+      page = currentPage
+    })
     it('should upload a select file', async () => {
       await expect(page).toUploadFile(
         'input[type="file"]',
